[MM-15977] Enhance mattermost version change behavior (#32)
* Enhance mattermost version change behavior

This change includes improvements to the ability of the operator
to move between mattermost versions. The operator will now run
a job of the requested image version in order to perform any
database migrations that may be required. If those complete
successfully then the deployment will be updated to use the new
mattermost image as well.

This change also improves logging and corrects an issue with `e2e`
tests not running the correct operator image.

* Update operator manifest